pyproject.toml
[tool.poetry]
name = "church_website_api"
version = "0.1.0"
description = "API для православных церквей."
authors = ["Alexander Dryannov <l3x.d@yandex.ru>"]
license = "GNU GPLv3"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.6"
djangorestframework = "^3.13.1"
Pillow = "^9.2.0"
django-cors-headers = "^3.13.0"
python-dotenv = "^0.20.0"
python-docx = "^0.8.11"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
